+++ src/md5/mod.rs
@@ -170,6 +170,27 @@ pub fn new() -> Update {
     Update::new()
 }
 
+/// Creates default hash instance.
+///
+/// Check [`Update`] for more details.
+///
+/// # Example
+///
+/// ```rust
+/// use chksum_hash::md5;
+///
+/// let digest = md5::default().update("data").digest();
+/// assert_eq!(
+///     digest.to_hex_lowercase(),
+///     "8d777f385d3dfec8815d20f7496026dc"
+/// );
+/// ```
+#[cfg_attr(all(release, feature = "inline"), inline)]
+#[must_use]
+pub fn default() -> Update {
+    Update::default()
+}
+
 /// Computes hash of given input.
 ///
 /// # Example
@@ -390,6 +411,13 @@ impl Update {
     }
 }
 
+impl Default for Update {
+    #[cfg_attr(all(release, feature = "inline"), inline)]
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 /// Represents finalized state.
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub struct Finalize {
@@ -422,6 +450,13 @@ impl Finalize {
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn test_default() {
+        let digest = default().digest();
+        assert_eq!(digest.to_hex_lowercase(), "d41d8cd98f00b204e9800998ecf8427e");
+        assert_eq!(verify(b"", digest), true);
+    }
 
     #[test]
     fn test_empty() {