+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `Default` trait to `State` and `Update` structs.
+
 ### Removed
 
 - Removed unnecessary `#[doc]` attributes for `Algorithm` variants.